pyproject.toml
[project]
name = "yandexcampbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Viktoriya Klyueva",email = "vika_klyueva@inbox.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot (>=22.4,<23.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
